Implement `par` task
The `par` task uses rayon to execute sub-tasks in parallel. If any of
the tasks fail, the `par` task will return the first error value it
encounters.

Making this change required converting `ExecutionError` from a type
alias to a struct in order to meet the `Send` trait bound rayon
requires.